9_vector_databases/6_lancedb/requirements.txt
lancedb 
python-dotenv 
langchain-openai 
langchain-community